+++ cli/CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.0] - 2026-08-12
+
+### Changed
+
+- Update for mcplease 0.3
+
 ## [0.2.1](https://github.com/jbr/mcplease/compare/mcplease-cli-v0.2.0...mcplease-cli-v0.2.1) - 2025-07-18
 
 ### Other